pyproject.toml
[project]
name = "nightcatwatermark"
version = "1.0.0"
description = "A desktop watermark tool with visible and blind watermark support"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "PyQt6>=6.6.0",
    "Pillow>=10.0.0",
    "blind-watermark>=0.4.4",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
]

[project.scripts]
nightcat = "main:main"

[project.gui-scripts]
nightcat-gui = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
